requirements.txt
py-trello>=0.16.0
pytz>=2020.1